% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basepop.R
\name{basepop}
\alias{basepop}
\title{Function to generate baseline population matrix.}
\usage{
basepop(
  inputsPath,
  regUAll,
  cur.scenario = "Constant_rate",
  scen.factor = 1,
  int.mig = FALSE
)
}
\arguments{
\item{inputsPath}{Input folder containing state subdirectories}

\item{regUAll}{Array containing state subfolder names}

\item{cur.scenario}{Current scenario rate (constant rate, etc.)}

\item{scen.factor}{Target population scenario}

\item{int.mig}{International migration rate (boolean, False = not applied)}
}
\value{
DataFrame containing baseline population matrix
}
\description{
Function to generate baseline population matrix.
}
